Added Automapper in App project, along with Import and Export DTOs. Fully implemented Orders API in the Client and Contributer controllers (partially in Designer too).

+++ CustomCADSolutions.Core/Models/CADModel.cs
@@ -23,7 +23,7 @@ namespace CustomCADSolutions.Core.Models
 
         public DateTime? CreationDate { get; set; }
         
-        public (short, short, short) Coords { get; set; }
+        public (int, int, int) Coords { get; set; }
 
         [RegularExpression(CadConstants.SpinAxisRegEx, ErrorMessage = CadConstants.SpinAxisErrorMessage)]
         public char? SpinAxis { get; set; }
